requirements.txt
blitz-bayesian-pytorch==0.2.7
numpy==1.19.4
pandas==1.1.4
statsmodels==0.12.1
torch==1.6.0
ray==1.0.1.post1
